chore(TimeUnit): Moved TimeUnit into nes-window-types, where we use it.
