+++ nes-data-types/src/API/CMakeLists.txt
@@ -13,6 +13,5 @@
 add_source_files(nes-data-types
         AttributeField.cpp
         Schema.cpp
-        TimeUnit.cpp
 )
