+++ nes-window-types/src/Measures/CMakeLists.txt
@@ -13,4 +13,5 @@
 add_source_files(nes-window-types
         TimeCharacteristic.cpp
         TimeMeasure.cpp
+        TimeUnit.cpp
 )
